requirements.txt


fastapi==0.120.1
uvicorn[standard]==0.38.0
python-multipart==0.0.20

sqlalchemy[asyncio]==2.0.36
aioodbc==0.5.0
pyodbc==5.2.0
alembic==1.15.0

python-jose[cryptography]==3.5.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.12.3
pydantic-settings==2.7.0

azure-communication-email==1.1.0
azure-identity==1.19.0

python-dateutil==2.9.0
pytz==2024.2

pyotp==2.9.0
qrcode[pil]==8.2

azure-storage-blob==12.27.1
